Refactor http utilities to common package
Moving several functions from `httphandler/utils.go` to a leaf package in `httputils/utils.go`.
These functions will be used in the next commit, that introduces the http catch-up service in the comm package.

No functional change.
