+++ internal/httphandler/db_request_handler_test.go
@@ -12,10 +12,9 @@ import (
 	"testing"
 	"time"
 
-	interrors "github.com/IBM-Blockchain/bcdb-server/internal/errors"
-
 	"github.com/IBM-Blockchain/bcdb-server/internal/bcdb"
 	"github.com/IBM-Blockchain/bcdb-server/internal/bcdb/mocks"
+	interrors "github.com/IBM-Blockchain/bcdb-server/internal/errors"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/constants"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/server/testutils"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/types"
