+++ internal/httphandler/utils.go
@@ -1,71 +1,27 @@
 // Copyright IBM Corp. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
+
 package httphandler
 
 import (
 	"encoding/base64"
 	"encoding/json"
 	"errors"
-	"fmt"
-	"log"
 	"net/http"
 	"strconv"
 	"time"
 
-	internalerror "github.com/IBM-Blockchain/bcdb-server/internal/errors"
-
-	"github.com/IBM-Blockchain/bcdb-server/internal/bcdb"
+	"github.com/IBM-Blockchain/bcdb-server/internal/httputils"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/constants"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/cryptoservice"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/types"
 	"github.com/gorilla/mux"
 )
 
-func MarshalOrPanic(response interface{}) []byte {
-	bytes, err := json.Marshal(response)
-	if err != nil {
-		panic(err)
-	}
-
-	return bytes
-}
-
-// SendHTTPResponse writes HTTP response back including HTTP code number and encode payload
-func SendHTTPResponse(w http.ResponseWriter, code int, payload interface{}) {
-	response, _ := json.Marshal(payload)
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(code)
-	if _, err := w.Write(response); err != nil {
-		log.Printf("Warning: failed to write response [%v] to the response writer\n", w)
-	}
-}
-
-type txHandler struct {
-	db bcdb.DB
-}
-
-// HandleTransaction handles transaction submission
-func (t *txHandler) handleTransaction(w http.ResponseWriter, tx interface{}, timeout time.Duration) {
-	// If timeout == 0, tx is async, otherwise it is synchronous.
-	resp, err := t.db.SubmitTransaction(tx, timeout)
-	if err != nil {
-		switch err.(type) {
-		case *internalerror.DuplicateTxIDError:
-			SendHTTPResponse(w, http.StatusBadRequest, &types.HttpResponseErr{ErrMsg: err.Error()})
-		case *internalerror.TimeoutErr:
-			SendHTTPResponse(w, http.StatusAccepted, &types.HttpResponseErr{ErrMsg: "Transaction processing timeout"})
-		default:
-			SendHTTPResponse(w, http.StatusInternalServerError, &types.HttpResponseErr{ErrMsg: err.Error()})
-		}
-		return
-	}
-	SendHTTPResponse(w, http.StatusOK, resp)
-}
-
 func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryType string, signVerifier *cryptoservice.SignatureVerifier) (interface{}, bool) {
 	querierUserID, signature, err := validateAndParseHeader(&r.Header)
 	if err != nil {
-		SendHTTPResponse(w, http.StatusBadRequest, &types.HttpResponseErr{ErrMsg: err.Error()})
+		httputils.SendHTTPResponse(w, http.StatusBadRequest, &types.HttpResponseErr{ErrMsg: err.Error()})
 		return nil, true
 	}
 
@@ -99,9 +55,9 @@ func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryTy
 			NodeId: params["nodeId"],
 		}
 	case constants.GetBlockHeader:
-		blockNum, err := getUintParam("blockId", params)
+		blockNum, err := httputils.GetUintParam("blockId", params)
 		if err != nil {
-			SendHTTPResponse(w, http.StatusBadRequest, err)
+			httputils.SendHTTPResponse(w, http.StatusBadRequest, err)
 			return nil, true
 		}
 
@@ -110,9 +66,9 @@ func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryTy
 			BlockNumber: blockNum,
 		}
 	case constants.GetPath:
-		startBlockNum, endBlockNum, err := getStartAndEndBlockNum(params)
+		startBlockNum, endBlockNum, err := httputils.GetStartAndEndBlockNum(params)
 		if err != nil {
-			SendHTTPResponse(w, http.StatusBadRequest, err)
+			httputils.SendHTTPResponse(w, http.StatusBadRequest, err)
 			return nil, true
 		}
 
@@ -122,9 +78,9 @@ func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryTy
 			EndBlockNumber:   endBlockNum,
 		}
 	case constants.GetTxProof:
-		blockNum, txIndex, err := getBlockNumAndTxIndex(params)
+		blockNum, txIndex, err := httputils.GetBlockNumAndTxIndex(params)
 		if err != nil {
-			SendHTTPResponse(w, http.StatusBadRequest, err)
+			httputils.SendHTTPResponse(w, http.StatusBadRequest, err)
 			return nil, true
 		}
 
@@ -134,9 +90,9 @@ func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryTy
 			TxIndex:     txIndex,
 		}
 	case constants.GetDataProof:
-		blockNum, err := getBlockNum(params)
+		blockNum, err := httputils.GetBlockNum(params)
 		if err != nil {
-			SendHTTPResponse(w, http.StatusBadRequest, err)
+			httputils.SendHTTPResponse(w, http.StatusBadRequest, err)
 			return nil, true
 		}
 
@@ -144,7 +100,7 @@ func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryTy
 		if _, ok := params["deleted"]; ok {
 			deleted, err = strconv.ParseBool(params["deleted"])
 			if err != nil {
-				SendHTTPResponse(w, http.StatusBadRequest, err)
+				httputils.SendHTTPResponse(w, http.StatusBadRequest, err)
 				return nil, true
 			}
 		}
@@ -162,15 +118,15 @@ func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryTy
 			TxId:   params["txId"],
 		}
 	case constants.GetHistoricalData:
-		version, err := getVersion(params)
+		version, err := httputils.GetVersion(params)
 		if err != nil {
-			SendHTTPResponse(w, http.StatusBadRequest, err)
+			httputils.SendHTTPResponse(w, http.StatusBadRequest, err)
 			return nil, true
 		}
 
 		v, isOnlyDeletesSet := params["onlydeletes"]
 		if isOnlyDeletesSet && v != "true" {
-			SendHTTPResponse(w, http.StatusBadRequest, &types.HttpResponseErr{
+			httputils.SendHTTPResponse(w, http.StatusBadRequest, &types.HttpResponseErr{
 				ErrMsg: "the onlydeletes parameters must be set only to 'true'",
 			})
 			return nil, true
@@ -220,9 +176,9 @@ func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryTy
 			TargetUserId: params["userId"],
 		}
 	case constants.GetMostRecentUserOrNode:
-		version, err := getVersion(params)
+		version, err := httputils.GetVersion(params)
 		if err != nil {
-			SendHTTPResponse(w, http.StatusBadRequest, err)
+			httputils.SendHTTPResponse(w, http.StatusBadRequest, err)
 			return nil, true
 		}
 
@@ -243,7 +199,7 @@ func extractVerifiedQueryPayload(w http.ResponseWriter, r *http.Request, queryTy
 
 	err, status := VerifyRequestSignature(signVerifier, querierUserID, signature, payload)
 	if err != nil {
-		SendHTTPResponse(w, status, err)
+		httputils.SendHTTPResponse(w, status, err)
 		return nil, true
 	}
 
@@ -304,82 +260,3 @@ func validateAndParseTxPostHeader(h *http.Header) (time.Duration, error) {
 	return timeout, nil
 }
 
-func getBlockNumAndTxIndex(params map[string]string) (uint64, uint64, error) {
-	blockNum, err := getUintParam("blockId", params)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	txIndex, err := getUintParam("idx", params)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	return blockNum, txIndex, nil
-}
-
-func getBlockNum(params map[string]string) (uint64, error) {
-	blockNum, err := getUintParam("blockId", params)
-	if err != nil {
-		return 0, err
-	}
-
-	return blockNum, nil
-}
-
-func getStartAndEndBlockNum(params map[string]string) (uint64, uint64, error) {
-	startBlockNum, err := getUintParam("startId", params)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	endBlockNum, err := getUintParam("endId", params)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	if endBlockNum < startBlockNum {
-		return 0, 0, &types.HttpResponseErr{
-			ErrMsg: fmt.Sprintf("query error: startId=%d > endId=%d", startBlockNum, endBlockNum),
-		}
-	}
-
-	return startBlockNum, endBlockNum, nil
-}
-
-func getVersion(params map[string]string) (*types.Version, error) {
-	if _, ok := params["blknum"]; !ok {
-		return nil, nil
-	}
-
-	blockNum, err := getUintParam("blknum", params)
-	if err != nil {
-		return nil, err
-	}
-
-	txNum, err := getUintParam("txnum", params)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.Version{
-		BlockNum: blockNum,
-		TxNum:    txNum,
-	}, nil
-}
-
-func getUintParam(key string, params map[string]string) (uint64, *types.HttpResponseErr) {
-	valStr, ok := params[key]
-	if !ok {
-		return 0, &types.HttpResponseErr{
-			ErrMsg: "query error - bad or missing literal: " + key,
-		}
-	}
-	val, err := strconv.ParseUint(valStr, 10, 64)
-	if err != nil {
-		return 0, &types.HttpResponseErr{
-			ErrMsg: "query error - bad or missing literal: " + key + " " + err.Error(),
-		}
-	}
-	return val, nil
-}
