+++ internal/httphandler/utils_test.go
@@ -3,60 +3,17 @@
 package httphandler
 
 import (
-	"encoding/json"
 	"errors"
 	"net/http"
-	"net/http/httptest"
 	"testing"
-
-	"github.com/IBM-Blockchain/bcdb-server/pkg/logger"
 
 	"github.com/IBM-Blockchain/bcdb-server/internal/bcdb/mocks"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/cryptoservice"
+	"github.com/IBM-Blockchain/bcdb-server/pkg/logger"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/server/testutils"
 	"github.com/IBM-Blockchain/bcdb-server/pkg/types"
-	"github.com/golang/protobuf/proto"
 	"github.com/stretchr/testify/require"
 )
-
-func TestSendHTTPResponse(t *testing.T) {
-	t.Parallel()
-
-	t.Run("ok status", func(t *testing.T) {
-		t.Parallel()
-
-		w := httptest.NewRecorder()
-		dbStatus := &types.GetDBStatusResponseEnvelope{
-			Response: &types.GetDBStatusResponse{
-				Header: &types.ResponseHeader{
-					NodeId: "testID",
-				},
-				Exist: false,
-			},
-		}
-		SendHTTPResponse(w, http.StatusOK, dbStatus)
-
-		require.Equal(t, http.StatusOK, w.Code)
-		actualDBStatus := &types.GetDBStatusResponseEnvelope{}
-		require.NoError(t, json.Unmarshal(w.Body.Bytes(), actualDBStatus))
-		require.True(t, proto.Equal(dbStatus, actualDBStatus))
-	})
-
-	t.Run("forbidden status", func(t *testing.T) {
-		t.Parallel()
-
-		w := httptest.NewRecorder()
-		err := &types.HttpResponseErr{
-			ErrMsg: "user does not have a read permission",
-		}
-		SendHTTPResponse(w, http.StatusForbidden, err)
-
-		require.Equal(t, http.StatusForbidden, w.Code)
-		actualErr := &types.HttpResponseErr{}
-		require.NoError(t, json.Unmarshal(w.Body.Bytes(), actualErr))
-		require.Equal(t, err, actualErr)
-	})
-}
 
 func TestVerifyRequestSignature(t *testing.T) {
 	lg, err := logger.New(&logger.Config{
